requirements-cpu.txt
advbox @ file:///D:/PrivacyProtectionSystem/AdvBox
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astor==0.8.1
asttokens==3.0.1
async-lru==2.3.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==26.1.0
babel==2.18.0
bce-python-sdk==0.9.68
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
crc32c==2.8
cryptography==46.0.7
cssselect==1.4.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
datasets==4.8.4
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.4
distro==1.9.0
easyocr==1.7.2
einops==0.8.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.136.0
fastjsonschema==2.21.2
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.47
gradio>=5.0.0,<6.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ImageIO==2.37.3
imagesize==2.0.0
imgaug==0.4.0
ipykernel==7.2.0
ipython==8.39.0
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lazy-loader==0.5
lmdb==2.2.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
modelscope==1.35.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.12.2
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.17.0
onnxoptimizer==0.3.13
openai==2.30.0
opencv-contrib-python==4.10.0.84
opencv-python==4.6.0.66
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==26.0
paddle2onnx==2.1.0
paddlefsl==1.1.0
paddlenlp==2.6.1
paddleocr==3.3.3
paddlepaddle==3.1.1
paddlex==3.3.13
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pdf2docx==0.5.12
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.4
polygraphy==0.49.26
premailer==3.10.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.2
psutil==7.2.2
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pypdfium2==5.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==4.1.0
pytz==2026.1.post1
pywinpty==3.0.3
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.3
rarfile==4.2
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==2.1.0
sentencepiece==0.2.1
seqeval==1.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.44.0
visualdl==2.5.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.7
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.23.0
